moved menu item context file to contexts folder
